Broke out Post Form
The post form is now its own reusable component.

The component <timeline-form> handles text and images for now. It could easily have videos and other things added to it. 

[(form)]="parentVariable" binds the values of the timeline-form's data to the parent's copy of it.

(submitData) is an event fired from within the component that tells the parent to submit the post using whichever method you define to take care of that.